+++ flaskProject/tests/test_ejemplo.py
@@ -1,31 +1,10 @@
-import os
-import tempfile
-
-import pytest
-
-from add_data import init_db
-from app import app
-from flask import request, jsonify
-
-@pytest.fixture
-def client():
-    db_fd, app.config['DATABASE'] = tempfile.mkstemp()
-    app.config['TESTING'] = True
-
-    with app.test_client() as client:
-        with app.app_context():
-            init_db()
-        yield client
-
-    os.close(db_fd)
-    os.unlink(app.config['DATABASE'])
-
+from tests.test_config import *
 
 
 
 
 def test_ejemplo():
-    # Se utiliza la keyword with porque es más seguro trabajar con recursos de esa
+    # Se utiliza la keyword "with" porque es más seguro trabajar con recursos de esa
     # manera. Al fin y al cabo es una asignación pero se asegura de una correcta
     # adquisición y liberación de recursos. Sinceramente no se muy bien como lo hace,
     # pero si teneis curiosidad podeis leeros esto de aquí:
@@ -34,7 +13,7 @@ def test_ejemplo():
     # lo equivalente sería un simple
     # c = app.test_client()
     with app.test_client() as c:
-        # Cuidado no llaméis a ninguna variable c o client.
+        # Cuidado no llaméis a ninguna variable "c" o "client".
         # Lo de c simplemente es pq es el nombre de la variable que estoy utilizando en los tests.
         # Lo de client es pq está la función de arriba que creo que es imprescindible que se llame así
         # y no se debería de cambiar.
@@ -42,7 +21,7 @@ def test_ejemplo():
         # Aquí ponemos el path del endpoint
         path = '/clients'
         # Aquí llamamos al enpoint en cuestión con el método y el path
-        r = c.open(path)
+        r = c.get(path)
         # Si la respuesta es un error o algo por el estilo esto revienta (No hay json)
         # Si teneis muchas dudas os recomiendo que pongais un breakpoint en una de estas
         # líneas y ejecuteis en modo debug. Os haréis una idea de que hay dentro y tal.
@@ -50,8 +29,10 @@ def test_ejemplo():
 
         assert "clients" in json_data.keys()
         assert r.status_code == 200
+
         clients = json_data["clients"]
         len_clients = len(clients)
+        #assert len_clients == 4
         client0 = clients[0]
 
         # Aquí estoy probando otras cosas para ver como funcionan
@@ -62,10 +43,12 @@ def test_ejemplo():
         r = c.open(path_get_client)
         json_data_get_client = r.get_json()
 
+        """
         # El delete no y eso me da muchas ganas de llorar.
         path_delete = "/client/1"
         r = c.delete(path_delete)
         json_data_delete = r.get_json()
+        """
 
         # Con esto quería comprobar que efectivamente el delete funciona
         # pero no lo hace :,>( . Por eso falla el test
@@ -109,13 +92,3 @@ def test_ejemplo2():
 
         assert "clients" in json_data.keys()
         assert r.status_code == 200
-
-
-
-
-
-
-
-
-
-
