+++ flaskProject/tests/tests_secundary/test_motos.py
@@ -3,6 +3,8 @@ import math
 import app
 from models.moto_model import MotoModel
 
+# Test
+
 m1 = MotoModel(
     state="ACTIVE",
     matricula="3333-MMM",
